remove inject gas & decrease gas to u64

+++ packages/vm/src/vm.rs
@@ -149,9 +149,9 @@ where
         .expect("Wasmer instance is not set. This is a bug in the lifecycle.")
     }
 
-    pub fn decrease_gas_left(&self, gas_used: u32) -> Result<(), Error> {
+    pub fn decrease_gas_left(&self, gas_used: u64) -> Result<(), Error> {
         let gas_left = self.get_gas_left();
-        if gas_used as u64 > gas_left {
+        if gas_used > gas_left {
             Err(Error::OutOfGasError)
         } else {
             self.set_gas_left(gas_left.saturating_sub(gas_used.into()));
